Refactorred `on` method 🕸️
